+++ edgy/core/db/querysets/base.py
@@ -174,7 +174,9 @@ class BaseQuerySet(
             offset = limit_offset
         self._offset = offset
 
-        self._select_related = set(select_related)
+        self._select_related: set[str] = set(select_related)
+        # computed only, replacable
+        self._select_related_weak: set[str] = set()
         self._prefetch_related = list(prefetch_related)
         self._batch_size = batch_size
         self._order_by: tuple[str, ...] = tuple(order_by)
@@ -254,6 +256,8 @@ class BaseQuerySet(
             extra_select=self._extra_select,
         )
         queryset.or_clauses.extend(self.or_clauses)
+        # copy anyway
+        queryset._select_related_weak.update(self._select_related_weak)
         queryset._cached_select_related_expression = self._cached_select_related_expression
         return cast("QuerySet", queryset)
 
@@ -288,15 +292,9 @@ class BaseQuerySet(
         # fetch all is in cache
         self._cache_fetch_all: bool = False
 
-    def _build_order_by_expression(self, order_by: Any, expression: Any) -> Any:
+    def _build_order_by_iterable(self, order_by: Any) -> Iterable:
         """Builds the order by expression"""
-        expression = expression.order_by(*(self._prepare_order_by(entry) for entry in order_by))
-        return expression
-
-    def _build_group_by_expression(self, group_by: Any, expression: Any) -> Any:
-        """Builds the group by expression"""
-        expression = expression.group_by(*(self._prepare_order_by(entry) for entry in group_by))
-        return expression
+        return (self._prepare_order_by(entry) for entry in order_by)
 
     async def build_where_clause(
         self, _: Any = None, tables_and_models: tables_and_models_type | None = None
@@ -444,7 +442,7 @@ class BaseQuerySet(
             ] = {}
 
             # Select related
-            for select_path in self._select_related:
+            for select_path in self._select_related.union(self._select_related_weak):
                 # For m2m relationships
                 model_class = self.model_class
                 former_table = maintable
@@ -671,16 +669,16 @@ class BaseQuerySet(
         expression = expression.where(await self.build_where_clause(self, tables_and_models))
 
         if self._order_by:
-            expression = self._build_order_by_expression(self._order_by, expression=expression)
+            expression = expression.order_by(*self._build_order_by_iterable(self._order_by))
+
+        if self._group_by:
+            expression = expression.group_by(*self._build_order_by_iterable(self._group_by))
 
         if self.limit_count:
             expression = expression.limit(self.limit_count)
 
         if self._offset:
             expression = expression.offset(self._offset)
-
-        if self._group_by:
-            expression = self._build_group_by_expression(self._group_by, expression=expression)
 
         if self.distinct_on is not None:
             expression = self._build_select_distinct(self.distinct_on, expression=expression)
@@ -798,18 +796,52 @@ class BaseQuerySet(
 
     def _prepare_order_by(self, order_by: str) -> Any:
         """
-        Prepares an order by expression from a string.
+        Prepares an order by or group by expression from a string.
 
         Args:
             order_by (str): The field name for ordering, optionally prefixed with '-' for descending.
 
         Returns:
-            Any: The SQLAlchemy column expression for ordering.
+            Any: The SQLAlchemy column expression for ordering and the select_path.
         """
         reverse = order_by.startswith("-")
         order_by = order_by.lstrip("-")
-        order_col = self.table.columns[order_by]
+        order_col, _ = clauses_mod.clean_field_to_column(
+            self.model_class,
+            field_path=order_by,
+            embed_parent=self.embed_parent_filters,
+            model_database=self.database,
+        )
         return order_col.desc() if reverse else order_col
+
+    def _update_select_related_order_by(self, order_by: tuple[str, ...], *, clear: bool) -> bool:
+        """
+        Extracts select related info
+
+        Args:
+            order_by (list[str]): The field name for ordering, optionally prefixed with '-' for descending.
+
+        """
+        related: set[str] = set()
+        for order_by_element in order_by:
+            order_by_element = order_by_element.lstrip("-")
+            result = clauses_mod.clean_field_to_column(
+                self.model_class,
+                field_path=order_by_element,
+                embed_parent=self.embed_parent_filters,
+                model_database=self.database,
+            )[1]
+            if result:
+                related.add(result)
+        if related and not self._select_related.union(self._select_related_weak).issuperset(
+            related
+        ):
+            self._cached_select_related_expression = None
+            if clear:
+                self._select_related_weak.clear()
+            self._select_related_weak.update(related)
+            return True
+        return False
 
     def _prepare_fields_for_distinct(self, distinct_on: str) -> sqlalchemy.Column:
         """
@@ -1586,6 +1618,8 @@ class QuerySet(BaseQuerySet):
         """
         queryset: QuerySet = self._clone()
         queryset._order_by = order_by
+        if queryset._update_select_related_order_by(order_by, clear=True):
+            queryset._update_select_related_order_by(queryset._group_by, clear=False)
         return queryset
 
     def reverse(self) -> QuerySet:
@@ -1638,6 +1672,8 @@ class QuerySet(BaseQuerySet):
         """
         queryset: QuerySet = self._clone()
         queryset._group_by = group_by
+        if queryset._update_select_related_order_by(group_by, clear=True):
+            queryset._update_select_related_order_by(queryset._order_by, clear=False)
         return queryset
 
     def distinct(self, first: bool | str = True, *distinct_on: str) -> QuerySet:
@@ -1716,7 +1752,7 @@ class QuerySet(BaseQuerySet):
         fields: Sequence[str] | str | None = None,
         exclude: Sequence[str] | set[str] = None,
         exclude_none: bool = False,
-    ) -> list[Any]:
+    ) -> list[dict]:
         """
         Returns the results in a python dictionary format.
         """
@@ -1724,19 +1760,17 @@ class QuerySet(BaseQuerySet):
         if isinstance(fields, str):
             fields = [fields]
 
-        rows: list[BaseModelType] = await self
-
         if fields is not None and not isinstance(fields, Iterable):
             raise QuerySetError(detail="Fields must be a suitable sequence of strings or unset.")
 
-        if not fields:
-            rows = [row.model_dump(exclude=exclude, exclude_none=exclude_none) for row in rows]
-        else:
-            rows = [
+        rows: list[BaseModelType] = await self
+        if fields:
+            return [
                 row.model_dump(exclude=exclude, exclude_none=exclude_none, include=fields)
                 for row in rows
             ]
-        return rows
+        else:
+            return [row.model_dump(exclude=exclude, exclude_none=exclude_none) for row in rows]
 
     async def values_list(
         self,
@@ -1748,6 +1782,8 @@ class QuerySet(BaseQuerySet):
         """
         Returns the results in a python dictionary format.
         """
+        if isinstance(fields, str):
+            fields = [fields]
         rows = await self.values(
             fields=fields,
             exclude=exclude,