+++ docs/release-notes.md
@@ -21,6 +21,10 @@ hide:
 - Fix issues with json-editor ajax and enums by simply inlining all schemas and disabling ajax.
 - For the json schema view: don't use ref_template, it isn't well supported and leads to problems with enums.
 - Fix crash when `None` is passed to CompositeField.
+- Fix values_list when using relation queryset.
+- Fix values_list when using only string.
+- Fix performance when using relation queryset.
+- Fix order_by when using relation queryset.
 
 ## 0.32.5